+++ Dockerfile
@@ -10,6 +10,8 @@ COPY package*.json ./
 
 RUN npm config set registry "https://registry.npmjs.org/"
 
+RUN npm cache clean --force
+
 RUN npm install -g npm@latest
 
 RUN npm ci
