+++ src/config/default.config.json
@@ -1,29 +1,36 @@
 {
-	"meta": {
-		"name": "family-chronicles-server",
-		"version": "0.0.1",
-		"description": "The Family Chronicles project is an attempt to develop a family tree management tool. The focus here is on an open API that should support existing standards, but also provide its own more complex interfaces. The vision is to make instances of the API linkable by administrators to form multiple family pedigrees into one big whole.",
-		"homepage_url": "https://github.com/Family-Chronicles/Family-Chronicles-Server#readme",
-		"author": {
-			"name": "Jonas Pfalzgraf",
-			"email": "info@josunlp.de"
-		},
-		"contributors": [
-			{
-				"name": "Jonas Pfalzgraf",
-				"email": "info@josunlp.de"
-			}
-		],
-		"license": "MPL-2.0",
-		"repository": "git+https://github.com/Family-Chronicles/Family-Chronicles-Server.git",
-		"homepage": "https://github.com/Family-Chronicles/Family-Chronicles-Server#readme",
-		"bugs": "https://github.com/Family-Chronicles/Family-Chronicles-Server/issues"
-	},
-	"database": {
-		"host": "mongodb://mongo:27017",
-		"databasename": "family-chronicles",
-		"username": "admin",
-		"password": "admin",
-		"collections": ["users", "data", "familys", "persons", "relations", "tags"]
-	}
+  "meta": {
+    "name": "family-chronicles-server",
+    "version": "0.0.1",
+    "description": "The Family Chronicles project is an attempt to develop a family tree management tool. The focus here is on an open API that should support existing standards, but also provide its own more complex interfaces. The vision is to make instances of the API linkable by administrators to form multiple family pedigrees into one big whole.",
+    "homepage_url": "https://github.com/Family-Chronicles/Family-Chronicles-Server#readme",
+    "author": {
+      "name": "Jonas Pfalzgraf",
+      "email": "info@josunlp.de"
+    },
+    "contributors": [
+      {
+        "name": "Jonas Pfalzgraf",
+        "email": "info@josunlp.de"
+      }
+    ],
+    "license": "MPL-2.0",
+    "repository": "git+https://github.com/Family-Chronicles/Family-Chronicles-Server.git",
+    "homepage": "https://github.com/Family-Chronicles/Family-Chronicles-Server#readme",
+    "bugs": "https://github.com/Family-Chronicles/Family-Chronicles-Server/issues"
+  },
+  "database": {
+    "host": "mongodb://mongo:27017",
+    "databasename": "family-chronicles",
+    "username": "admin",
+    "password": "admin",
+    "collections": [
+      "users",
+      "data",
+      "familys",
+      "persons",
+      "relations",
+      "tags"
+    ]
+  }
 }
